Changed t to int

+++ Graphs/Pontes num Grafo.cpp
@@ -1,7 +1,7 @@
 //SE TIRA-LAS O GRAFO FICA DESCONEXO
 // OBS: PRESTAR ATENCAO EM SELF-LOOPS, É MELHOR NÃO ADICIONA-LOS
 // SO FUNCIONA EM GRAFO NÃO DIRECIONADO
-float t=1;
+int t=1;
 vector<int> T((int)2e6,0); //Tempo necessário para chegar naquele vértice na dfs 
 vector<int> adj[(int)2e6]; 
 vector<int> Low((int)2e6); // Tempo “mínimo” para chegar naquele vértice na dfs
